Fixed a bug in the client where the chat window was being destroyed too soon. Just before the window is closed, a part command is issued, then the close event is ended. When the client receives the notification from the server that it has parted, then the window is closed and destroyed.
Added a 'welcome' message on successful ident and changed the client to wait until it receives that before requesting the active room list.

Jiggled around the wshandler/client calls. Added simplechat_client:proxy_call/2 which makes calls to the client, and tags the result with the call that created it.